fix: photoTools에 무드 배경화면 기능 모듈화 1. 기록 카드에 사진이 없는 경우 무드 배경화면을 줘야하므로 모듈화 시킴
