+++ lib/photoTools.js
@@ -1,6 +1,6 @@
 import fetch from "node-fetch";
 import './env.js'
-
+import supabaseAdmin from './supabaseAdmin.js';
 
 // 여러 개의 장소 ID에 대해 사진 이름을 비동기적으로 가져오는 함수
 async function _getPhotoNames(...spaceIds) {
@@ -73,6 +73,77 @@ async function getPhotoUrls(...spaceIds) {
     return ret;
 }
 
+
+async function getMoodWallpaper(mood_id) {
+    if(mood_id.length === 0){
+        const { data:url, error } = await supabaseAdmin
+            .storage
+            .from('wallpaper')
+            .createSignedUrl('general/Rectangle 34627910.png', 60)
+
+        if (error) {
+            console.error('Wallpaper URL 생성 실패:', error);
+            return res.status(500).json({ error: 'Wallpaper URL 생성 실패' });
+        }
+        return {url: url.signedUrl, error: null};
+    }
+
+    const { data:mood_tags_data, error:mood_tags_error } = await supabaseAdmin
+        .from('mood_tags')
+        .select('*')
+    if (mood_tags_error) {
+        console.error('Mood tags 조회 실패:', mood_tags_error);
+        return { error: new Error('Mood tags 조회 실패') };
+    }
+    console.log('mood_tags_data:', mood_tags_data);
+
+    const kr_to_en = Object.fromEntries(mood_tags_data.map(tag => [tag.mood_id.trim(), tag.tag_en.trim()]));
+    console.log('kr_to_en:', kr_to_en);
+    const mood_id_en = mood_id.map(id => kr_to_en[id] || id)
+    console.log('mood_id_en:', mood_id_en);
+    let wallpaper_name = []
+
+    for(const tag of mood_id_en){
+        const { data, error } = await supabaseAdmin
+            .storage
+            .from('wallpaper')
+            .list(tag, {
+                limit: 100,
+                offset: 0,
+                sortBy: { column: 'name', order: 'asc' },
+            })
+        if (error) {
+            console.error('Wallpaper 목록 조회 실패:', error);
+            return { error: new Error('Wallpaper 목록 조회 실패') };
+        }
+        const names = data.map(item => `${tag}/${item.name}`);
+        wallpaper_name.push(...names);
+    }
+    console.log('선택된 Wallpaper 후보:', wallpaper_name);
+    if(wallpaper_name.length === 0){
+        const { data:url, error } = await supabaseAdmin
+            .storage
+            .from('wallpaper')
+            .createSignedUrl('general/Rectangle 34627910.png', 60)
+            
+        if (error) {
+            console.error('Wallpaper URL 생성 실패:', error);
+            return { error: new Error('Wallpaper URL 생성 실패') };
+        }
+        return {url: url.signedUrl, error: null};
+    }   
+    // 랜덤으로 하나 전송
+    const randomWallpaper = wallpaper_name[Math.floor(Math.random() * wallpaper_name.length)];
+    console.log('선택된 랜덤 Wallpaper:', randomWallpaper);
+    const { data: signedUrl, error: urlError } = await supabaseAdmin.storage.from('wallpaper').createSignedUrl(randomWallpaper, 60);
+    if (urlError) {
+        console.error('Wallpaper URL 생성 실패:', urlError);
+        return { error: new Error('Wallpaper URL 생성 실패') };
+    }
+    return {url: signedUrl.signedUrl, error: null};
+}
+
+// getMoodWallpaper(["활기찬"]).then(console.log).catch(console.error);
 // await getPhotoUrls("ChIJOUwYKJXvYTURRNQvZGGgndk", "ChIJ8-S81GrvYTURcGwu8ujd-i0", "ChIJm8OaS2fvYTURSnkXD7MLS9E", "ChIJ0ZX8TGfvYTURyeQbiu_HI5M").then(console.log).catch(console.error);
 
-export default {getPhotoUrls};
+export default {getPhotoUrls, getMoodWallpaper};
